--- Folder_20230624_181343/media/txt/output.txt
@@ -1,1 +1,0 @@
- Bonjour, c'est juste un test pour savoir d'efficacité de cette application.
