--- Folder_20230624_181351/media/txt/output.txt
@@ -1,1 +1,0 @@
-through sesatic patients pillars les actions de notre premier ju àbye ce
