--- Folder_20230624_181505/media/txt/output.txt
@@ -1,1 +1,0 @@
- Okay, see 304 microphone tests some about a 2.5 feet away from the microphone and see 304, the echo we just room in the company. This is the sound check with the idea of using it for echo testing for the convulver it's a look for a package.
