+++ requirements.txt
@@ -3,7 +3,7 @@ moviepy
 # ffmpeg
 ffmpeg-python
 setuptools-rust
-tiktoken
+tiktoken==0.3.3
 # ffmpeg-python==0.2.0
 streamlit
 datetime
@@ -15,3 +15,5 @@ mutagen
 pydub
 yt_dlp
 git+https://github.com/openai/whisper.git 
+python-dotenv
+azure-ai-textanalytics
